Убрал асинхронность из приложения
