+++ config/settings.py
@@ -122,16 +122,6 @@ STATIC_URL = 'static/'
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 
-# Celery Configuration Options
-CELERY_TIMEZONE = TIME_ZONE
-CELERY_TASK_TRACK_STARTED = True
-CELERY_TASK_TIME_LIMIT = 30 * 60
-
-CELERY_BROKER_URL = os.getenv('CELERY_BROKER_URL')
-
-CELERY_RESULT_BACKEND = os.getenv('CELERY_RESULT_BACKEND')
-
-
 # Настройки SMTP-сервера
 EMAIL_BACKEND = os.getenv('EMAIL_BACKEND')
 EMAIL_HOST = os.getenv('EMAIL_HOST')
